- These are continued updates for #1193 - Fixed another bug in SinkDetect that had to do with the stopping condition for flagging trails down the row, and removed copy error
git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/hstcal/trunk@41640 fe389314-cf27-0410-b35b-8c050e845b92

+++ pkg/wfc3/calwf3/wf3cte/wf3cte.c
@@ -1380,7 +1380,7 @@ int sim_colreadout_l(double *pixi, double *pixo, double *pixf, CTEParams *cte){
                     if ( pix_1 >= cte->qlevq_data[w] ){
                         prem_3 = ( cte->dpdew_data[w] / (double)cte->n_par) * pixf[j];                        
                         if (ttrap < cte->cte_len)
-                            padd_3 = (cprof->data,w,ttrap)*ftrap;
+                            padd_3 = (double)(cprof->data,w,ttrap)*ftrap;
                         ttrap=0;
                         ftrap=prem_3;
                     }
